src/reference/fancy_poc/multimodal_chat_robot/src/multimodal_chat_robot_server/requirements.txt
accelerate==0.17.1
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
brotlipy==0.7.0
cached-property==1.5.2
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
datasets==2.11.0
diffusers==0.15.0
dill==0.3.6
ffmpeg-python==0.2.0
filelock==3.10.7
flatbuffers==1.12
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gensim==4.3.1
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
jieba==0.42.1
Jinja2==3.1.2
joblib==1.1.1
keras-nightly==2.5.0.dev2021032900
keyboard==0.13.5
kiwisolver==1.4.4
llvmlite==0.39.1
loguru==0.6.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai==0.27.2
openai-whisper==20230314
opencv-contrib-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==8.3.2
pip==23.0.1
pooch==1.4.0
psutil==5.9.4
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
responses==0.18.0
rsa==4.7.2
scikit-learn==1.2.1
scipy==1.10.0
setuptools==65.6.3
six==1.16.0
smart-open==6.3.0
soundfile==0.12.1
sympy==1.11.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
text2vec==1.1.7
threadpoolctl==2.2.0
tiktoken==0.3.1
tokenizers==0.13.2
torch==1.13.1+cu117
tqdm==4.65.0
transformers==4.27.4
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==0.16.0
wheel==0.38.4
win-inet-pton==1.1.0
win32-setctime==1.1.0
wincertstore==0.2
wrapt==1.12.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
